Constructed setup class for manufactured solution
